Address comments - cleanup unused firbase infra, move interface files, etc
