--- .firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "shareable-mvp"
-  }
-}
