+++ package.json
@@ -18,7 +18,6 @@
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
     "@vitejs/plugin-react": "^3.0.0",
-    "firebase-tools": "^11.21.0",
     "typescript": "^4.9.3",
     "vite": "^4.0.0"
   }
