+++ src/actions/filedRequestActions.ts
@@ -1,9 +1,9 @@
 import { firestore } from '../firebase'
-import {FiledRequest, REQUESTS_COLLECTION} from "../types/filedRequest";
+import {FiledRequest, REQUESTS_COLLECTION} from "../interfaces/filedRequest";
 import {v4 as uuidv4} from 'uuid';
 import {doc, setDoc} from "firebase/firestore";
-import {ORGANIZATIONS_COLLECTION, TEST_ORGANIZATION} from "../types/organization";
-import {GROUPS_COLLECTION, TEST_GROUP} from "../types/group";
+import {ORGANIZATIONS_COLLECTION, TEST_ORGANIZATION} from "../interfaces/organization";
+import {GROUPS_COLLECTION, TEST_GROUP} from "../interfaces/group";
 
 export const addFiledRequest = (filedRequest: FiledRequest) => {
     const uuid = uuidv4()
