+++ src/actions/userAuthActions.ts
@@ -1,21 +1,25 @@
 import { firebaseApp, firestore } from '../firebase'
 import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, GoogleAuthProvider, signInWithPopup, onAuthStateChanged } from "firebase/auth";
 import { doc, setDoc } from "firebase/firestore";
-import {TEST_ORGANIZATION} from "../types/organization";
-import {TEST_GROUP} from "../types/group";
+import {TEST_ORGANIZATION} from "../interfaces/organization";
+import {TEST_GROUP} from "../interfaces/group";
+import {UserCredential} from "@firebase/auth";
+import {UserEmailCredentials, USERS_COLLECTION} from "../interfaces/user";
 
 const auth = getAuth(firebaseApp);
 
-export const signUp = (userCredentials: any) => {
-    createUserWithEmailAndPassword(auth, userCredentials.email, userCredentials.password)
-        .then((userCredential) => {
-           setDoc(doc(firestore, "users", userCredential.user.uid), {
+export const signUp = (userEmailCredentials: UserEmailCredentials) => {
+    const {email, password} = userEmailCredentials;
+    createUserWithEmailAndPassword(auth, email, password)
+        .then((firebaseUserCredential: UserCredential) => {
+            const userId: String = firebaseUserCredential.user.uid;
+            setDoc(doc(firestore, USERS_COLLECTION, userId), {
                organizationId: [TEST_ORGANIZATION],
                groupsIds: [TEST_GROUP]
-           }).then(() => {
+            }).then(() => {
 
-           });
-         })
+            });
+        })
         .catch((error) => {
             const errorCode = error.code;
             const errorMessage = error.message;
@@ -23,11 +27,12 @@ export const signUp = (userCredentials: any) => {
         });
 }
 
-export const signIn = (userCredentials: any) => {
-    signInWithEmailAndPassword(auth, userCredentials.username, userCredentials.password)
-        .then((userCredential) => {
+export const signIn = (userEmailCredentials: UserEmailCredentials) => {
+    const {email, password} = userEmailCredentials;
+    signInWithEmailAndPassword(auth, email, password)
+        .then((firebaseUserCredential: UserCredential) => {
             // Signed in
-            const user = userCredential.user;
+            const user = firebaseUserCredential.user;
             // ...
         })
         .catch((error) => {
@@ -39,8 +44,8 @@ export const signIn = (userCredentials: any) => {
 export const signInWithGoogle = () => {
     const googleAuthProvider: GoogleAuthProvider = new GoogleAuthProvider();
     signInWithPopup(auth, googleAuthProvider)
-        .then((result) => {
-            const user = result.user;
+        .then((firebaseUserCredential: UserCredential) => {
+            const user = firebaseUserCredential.user;
             // ...
         }).catch((error) => {
         // Handle Errors here.
@@ -56,5 +61,3 @@ export const signOut = () => {
         console.log("Signed out")
     });
 }
-
-
