+++ src/interfaces/insight.ts
