+++ src/interfaces/user.ts
@@ -1,0 +1,12 @@
+export const USERS_COLLECTION = "users"
+export interface User {
+    organizationId: string;
+    groupsIds: string[]
+
+    // Other misc user metadata
+}
+
+export interface UserEmailCredentials {
+    email: String;
+    password: String;
+}
