--- src/types/user.ts
@@ -1,6 +1,0 @@
-export interface User {
-    organizationId: string;
-    groupsIds: string[]
-
-    // Other misc user metadata
-}
